TelegramDocBot/requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# OpenAI (Whisper + GPT)
openai==1.12.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# Document creation
python-docx==1.1.0

# Environment variables
python-dotenv==1.0.1

# Async support
aiohttp==3.9.3
